refactor FileManager, add module into LexicalAnalyzer
